+++ CHANGES.md
@@ -1,4 +1,13 @@
+# Changes
+
 ## Pending
+
+* Install `flexdll[_initer]_msvc[64].obj` to `bin/` alongside existing
+  `flexlink.exe` so that flexlink can run standalone without setting
+  FLEXDIR environment variable. Bug report at
+  <https://github.com/diskuv/dkml-installer-ocaml/issues/40>
+
+## `4.14.0~v1.2.0`
 
 * Add `/DEBUG:FULL` to MSVC linker and `-Zi -Zd` to MSVC assembler, plus
   existing `-Z7` in MSVC compiler, when `dkml-option-debuginfo` is present